DECA-193: Ensures that config keys are not unicode.
When switching to the latest version of Cherrypy (3.2.2), which is needed to properly handle dynamic urls, there is a problem when the config keys are already unicode values. Cherrypy now attempts to convert them to unicode but will fail if they already are. This will just encode the keys as utf-8 ahead of time. If/when this is fixed by cherrypy, this encoding step can be removed.