DECA-193: switched to the MethodDispatcher
The method dispatcher is more specific to a RESTful approach, as the method names are based on the standard REST methods (GET, PUT, POST, DELETE). These python methods will be called automatically based on the HTTP method used.

This change though, means that there will be a lot of updating needed. For example we'll have to change the methods for every controller to comply with this scheme. Also, with Cherrypy 3.2.2 we can use _cp_dispatch to better handle dynamic URLS.